Fix typos in SETUP-DEVICE.md, give context
I first encountered this article on the internet when looking for other information about IoT Edge preconfigured OSes.

Noticed a couple spelling mistakes and inconsistencies, those I fixed here. (Please note I do not have this starter kit, and so was not able to try these instructions - I can't vouch for correctness at all.)

Also, the article states as matter-of-fact that the Raspberry Pi has a preconfigured IP of 192.168.4.1. Important to clarify that is (as far as I can tell) specific to the Grover Starter kit - that does not apply to Raspberry Pis at-large.